+++ frontend-maven-plugin/src/main/java/com/github/eirslett/maven/plugins/frontend/mojo/YarnMojo.java
@@ -1,8 +1,8 @@
 package com.github.eirslett.maven.plugins.frontend.mojo;
 
-import com.github.eirslett.maven.plugins.frontend.lib.FrontendPluginFactory;
-import com.github.eirslett.maven.plugins.frontend.lib.ProxyConfig;
-import com.github.eirslett.maven.plugins.frontend.lib.TaskRunnerException;
+import java.io.File;
+import java.util.Collections;
+
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugins.annotations.Component;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
@@ -11,29 +11,31 @@ import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.settings.crypto.SettingsDecrypter;
 import org.sonatype.plexus.build.incremental.BuildContext;
 
-import java.io.File;
-import java.util.Collections;
+import com.github.eirslett.maven.plugins.frontend.lib.FrontendPluginFactory;
+import com.github.eirslett.maven.plugins.frontend.lib.ProxyConfig;
+import com.github.eirslett.maven.plugins.frontend.lib.TaskRunnerException;
 
-@Mojo(name="npm",  defaultPhase = LifecyclePhase.GENERATE_RESOURCES)
-public final class NpmMojo extends AbstractFrontendMojo {
+@Mojo(name = "yarn", defaultPhase = LifecyclePhase.GENERATE_RESOURCES)
+public final class YarnMojo extends AbstractFrontendMojo {
 
     private static final String NPM_REGISTRY_URL = "npmRegistryURL";
-    
+
     /**
      * npm arguments. Default is "install".
      */
-    @Parameter(defaultValue = "install", property = "frontend.npm.arguments", required = false)
+    @Parameter(defaultValue = "", property = "frontend.yarn.arguments", required = false)
     private String arguments;
 
-    @Parameter(property = "frontend.npm.npmInheritsProxyConfigFromMaven", required = false, defaultValue = "true")
-    private boolean npmInheritsProxyConfigFromMaven;
+    @Parameter(property = "frontend.yarn.yarnInheritsProxyConfigFromMaven", required = false,
+        defaultValue = "true")
+    private boolean yarnInheritsProxyConfigFromMaven;
 
     /**
      * Registry override, passed as the registry option during npm install if set.
      */
     @Parameter(property = NPM_REGISTRY_URL, required = false, defaultValue = "")
     private String npmRegistryURL;
-    
+
     @Parameter(property = "session", defaultValue = "${session}", readonly = true)
     private MavenSession session;
 
@@ -46,7 +48,7 @@ public final class NpmMojo extends AbstractFrontendMojo {
     /**
      * Skips execution of this mojo.
      */
-    @Parameter(property = "skip.npm", defaultValue = "${skip.npm}")
+    @Parameter(property = "skip.yarn", defaultValue = "${skip.yarn}")
     private boolean skip;
 
     @Override
@@ -56,26 +58,28 @@ public final class NpmMojo extends AbstractFrontendMojo {
 
     @Override
     public void execute(FrontendPluginFactory factory) throws TaskRunnerException {
-        File packageJson = new File(workingDirectory, "package.json");
-        if (buildContext == null || buildContext.hasDelta(packageJson) || !buildContext.isIncremental()) {
+        File packageJson = new File(this.workingDirectory, "package.json");
+        if (this.buildContext == null || this.buildContext.hasDelta(packageJson)
+            || !this.buildContext.isIncremental()) {
             ProxyConfig proxyConfig = getProxyConfig();
-            factory.getNpmRunner(proxyConfig, getRegistryUrl()).execute(arguments, environmentVariables);
+            factory.getYarnRunner(proxyConfig, getRegistryUrl()).execute(this.arguments,
+                this.environmentVariables);
         } else {
-            getLog().info("Skipping npm install as package.json unchanged");
+            getLog().info("Skipping yarn install as package.json unchanged");
         }
     }
 
     private ProxyConfig getProxyConfig() {
-        if (npmInheritsProxyConfigFromMaven) {
-            return MojoUtils.getProxyConfig(session, decrypter);
+        if (this.yarnInheritsProxyConfigFromMaven) {
+            return MojoUtils.getProxyConfig(this.session, this.decrypter);
         } else {
-            getLog().info("npm not inheriting proxy config from Maven");
+            getLog().info("yarn not inheriting proxy config from Maven");
             return new ProxyConfig(Collections.<ProxyConfig.Proxy>emptyList());
         }
     }
 
     private String getRegistryUrl() {
         // check to see if overridden via `-D`, otherwise fallback to pom value
-        return System.getProperty(NPM_REGISTRY_URL, npmRegistryURL);
+        return System.getProperty(NPM_REGISTRY_URL, this.npmRegistryURL);
     }
 }