+++ frontend-plugin-core/src/main/java/com/github/eirslett/maven/plugins/frontend/lib/NodeExecutorConfig.java
@@ -5,15 +5,16 @@ import java.io.File;
 public interface NodeExecutorConfig {
   File getNodePath();
   File getNpmPath();
+  File getInstallDirectory();
   File getWorkingDirectory();
   Platform getPlatform();
 }
 
 final class InstallNodeExecutorConfig implements NodeExecutorConfig {
 
-  private static final String NODE_WINDOWS = "\\node\\node.exe";
-  private static final String NODE_DEFAULT = "/node/node";
-  private static final String NPM = "/node/node_modules/npm/bin/npm-cli.js";
+  private static final String NODE_WINDOWS = NodeInstaller.INSTALL_PATH.replaceAll("/", "\\\\") + "\\node.exe";
+  private static final String NODE_DEFAULT = NodeInstaller.INSTALL_PATH + "/node";
+  private static final String NPM = NodeInstaller.INSTALL_PATH + "/node_modules/npm/bin/npm-cli.js";
 
   private final InstallConfig installConfig;
 
@@ -32,7 +33,11 @@ final class InstallNodeExecutorConfig implements NodeExecutorConfig {
     return new File(installConfig.getInstallDirectory() + Utils.normalize(NPM));
   }
 
-
+  @Override
+  public File getInstallDirectory() {
+    return installConfig.getInstallDirectory();
+  }
+  
   @Override
   public File getWorkingDirectory() {
     return installConfig.getWorkingDirectory();