+++ frontend-plugin-core/src/main/java/com/github/eirslett/maven/plugins/frontend/lib/YarnTaskExecutor.java
@@ -1,47 +1,47 @@
 package com.github.eirslett.maven.plugins.frontend.lib;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import static com.github.eirslett.maven.plugins.frontend.lib.Utils.implode;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-
-import static com.github.eirslett.maven.plugins.frontend.lib.Utils.implode;
-import static com.github.eirslett.maven.plugins.frontend.lib.Utils.normalize;
-import static com.github.eirslett.maven.plugins.frontend.lib.Utils.prepend;
 import java.util.Map;
 
-abstract class NodeTaskExecutor {
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+abstract class YarnTaskExecutor {
     private static final String DS = "//";
+
     private static final String AT = "@";
 
     private final Logger logger;
-    private final String taskName;
-    private final String taskLocation;
-    private final List<String> additionalArguments;
-    private final NodeExecutorConfig config;
 
-    public NodeTaskExecutor(NodeExecutorConfig config, String taskLocation) {
+    private final String taskName;
+
+    private final List<String> additionalArguments;
+
+    private final YarnExecutorConfig config;
+
+    public YarnTaskExecutor(YarnExecutorConfig config, String taskLocation) {
         this(config, taskLocation, Collections.<String> emptyList());
     }
 
-    public NodeTaskExecutor(NodeExecutorConfig config, String taskName, String taskLocation) {
+    public YarnTaskExecutor(YarnExecutorConfig config, String taskName, String taskLocation) {
         this(config, taskName, taskLocation, Collections.<String> emptyList());
     }
 
-    public NodeTaskExecutor(NodeExecutorConfig config, String taskLocation, List<String> additionalArguments) {
+    public YarnTaskExecutor(YarnExecutorConfig config, String taskLocation,
+        List<String> additionalArguments) {
         this(config, getTaskNameFromLocation(taskLocation), taskLocation, additionalArguments);
     }
 
-    public NodeTaskExecutor(NodeExecutorConfig config, String taskName, String taskLocation,
-            List<String> additionalArguments) {
-        this.logger = LoggerFactory.getLogger(getClass());
+    public YarnTaskExecutor(YarnExecutorConfig config, String taskName, String taskLocation,
+        List<String> additionalArguments) {
+        logger = LoggerFactory.getLogger(getClass());
         this.config = config;
         this.taskName = taskName;
-        this.taskLocation = taskLocation;
         this.additionalArguments = additionalArguments;
     }
 
@@ -51,40 +51,22 @@ abstract class NodeTaskExecutor {
 
     public final void execute(String args, Map<String, String> environment) throws TaskRunnerException {
         final List<String> arguments = getArguments(args);
-        execute(arguments, environment);
-    }
-
-    public final void execute(List<String> arguments, Map<String, String> environment) throws TaskRunnerException {
-        final String absoluteTaskLocation = getAbsoluteTaskLocation();
         logger.info("Running " + taskToString(taskName, arguments) + " in " + config.getWorkingDirectory());
 
         try {
-            final int result = new NodeExecutor(config,
-                    prepend(absoluteTaskLocation, arguments),
-                    environment).executeAndRedirectOutput(logger);
+            final int result =
+                new YarnExecutor(config, arguments, environment).executeAndRedirectOutput(logger);
             if (result != 0) {
                 throw new TaskRunnerException(
-                        taskToString(taskName, arguments) + " failed. (error code " + result + ")", result);
+                    taskToString(taskName, arguments) + " failed. (error code " + result + ")");
             }
         } catch (ProcessExecutionException e) {
-            throw new TaskRunnerException(taskToString(taskName, arguments) + " failed.", e, e.getExitCode());
+            throw new TaskRunnerException(taskToString(taskName, arguments) + " failed.", e);
         }
-    }
-
-    private String getAbsoluteTaskLocation() {
-        String location = normalize(taskLocation);
-        if (Utils.isRelative(taskLocation)) {
-            File taskFile = new File(config.getWorkingDirectory(), location);
-            if (!taskFile.exists()) {
-                taskFile = new File(config.getInstallDirectory(), location);
-            }
-            location = taskFile.getAbsolutePath();
-        }
-        return location;
     }
 
     private List<String> getArguments(String args) {
-        List<String> arguments = new ArrayList<String>();
+        List<String> arguments = new ArrayList<>();
         if (args != null && !args.equals("null") && !args.isEmpty()) {
             arguments.addAll(Arrays.asList(args.split("\\s+")));
         }
@@ -98,7 +80,7 @@ abstract class NodeTaskExecutor {
     }
 
     private static String taskToString(String taskName, List<String> arguments) {
-        List<String> clonedArguments = new ArrayList<String>(arguments);
+        List<String> clonedArguments = new ArrayList<>(arguments);
         for (int i = 0; i < clonedArguments.size(); i++) {
             final String s = clonedArguments.get(i);
             final boolean maskMavenProxyPassword = s.contains("proxy=");
@@ -121,13 +103,15 @@ abstract class NodeTaskExecutor {
                 final int lastAtCharIndex = proxyString.lastIndexOf(AT);
                 boolean hasPossibleURIUserInfo = firstDoubleSlashIndex < lastAtCharIndex;
                 if (hasPossibleURIUserInfo) {
-                    final String userInfo = proxyString.substring(firstDoubleSlashIndex + DS.length(), lastAtCharIndex);
+                    final String userInfo =
+                        proxyString.substring(firstDoubleSlashIndex + DS.length(), lastAtCharIndex);
                     final String[] userParts = userInfo.split(":");
                     if (userParts.length > 0) {
                         final int startOfUserNameIndex = firstDoubleSlashIndex + DS.length();
-                        final int firstColonInUsernameOrEndOfUserNameIndex = startOfUserNameIndex
-                                + userParts[0].length();
-                        final String leftPart = proxyString.substring(0, firstColonInUsernameOrEndOfUserNameIndex);
+                        final int firstColonInUsernameOrEndOfUserNameIndex =
+                            startOfUserNameIndex + userParts[0].length();
+                        final String leftPart =
+                            proxyString.substring(0, firstColonInUsernameOrEndOfUserNameIndex);
                         final String rightPart = proxyString.substring(lastAtCharIndex);
                         retVal = leftPart + ":***" + rightPart;
                     }