+++ CHANGELOG.md
@@ -1,10 +1,28 @@
 # frontend-maven-plugin
 
+Last public release: [![Maven Central](https://maven-badges.herokuapp.com/maven-central/com.github.eirslett/frontend-maven-plugin/badge.svg?style=flat)](https://maven-badges.herokuapp.com/maven-central/com.github.eirslett/frontend-maven-plugin/)
+
 ## Changelog
 
-### 1.2-SNAPSHOT
+### 1.4
 
-* 
+* Add maven.frontend.failOnError and maven.test.failure.ignore flags to best manage integration-test
+* Fix #41: Replaced ProcessBuilder usage with commons-exec
+* Use InstallDirectory to locate node tasks instead of the WorkingDirectory as fallback
+* Fix 531: update lifecycle-mapping-metadata.xml for yarn
+* Fix execute goal for gulp
+* Fix #532: fix NullPointerException for invalid yarn version
+* Added bower proxy ignore parameter
+* Document how to skip package managers and build tools'
+
+### 1.3
+
+* Fix `yarn` for Windows
+* Fix #515: Change "yarn warning" from ERROR to WARNING in log
+
+### 1.2
+
+* New goals `yarn` and `install-node-and-yarn` for Yarn support 
 
 ### 1.1